requirements_backup.txt
python-dotenv==1.2.2
unstructured==0.22.22
langchain-core==1.3.0
langchain-community==0.4.1
langchain-text-splitters==1.1.2
langchain-chroma==1.1.0
langchain-google-genai==4.2.2
langchain-ollama==1.1.0
langchain-experimental==0.4.1
